Implement terminal box with buttons.
Install bootstrap.
Remov unecessary css.
Ensur main app is full screen.
Implement terminalBox, a contained that controls the terminalText and terminalInput and contains terminalMenuBar.
Implement terminalMenuBar as stub.
Implement terminalInput as a group of buttons that will display text in terminalText.
Implement terminalText as a way of displaying the printouts.
Implement prinouts.json, a list of commands for the user to enter, and messages for terminalText to display.

Currently the display text is string only, e.g. there are no links.